+++ Havok.Physics/HavokTriggerEvent.cs
@@ -1,5 +1,4 @@
 using Havok.Physics;
-using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
 
 namespace Unity.Physics
@@ -11,33 +10,27 @@ namespace Unity.Physics
         [NativeDisableUnsafePtrRestriction]
         private readonly HpLinkedRange* m_EventDataRange;
 
-        private readonly NativeSlice<RigidBody> m_Bodies;
-
-        internal HavokTriggerEvents(HpLinkedRange* eventDataRange, NativeSlice<RigidBody> bodies)
+        internal HavokTriggerEvents(HpLinkedRange* eventDataRange)
         {
             m_EventDataRange = eventDataRange;
-            m_Bodies = bodies;
         }
 
         public Enumerator GetEnumerator()
         {
-            return new Enumerator(m_EventDataRange, m_Bodies);
+            return new Enumerator(m_EventDataRange);
         }
 
         public struct Enumerator /* : IEnumerator<TriggerEvent> */
         {
             private HpLinkedRange* m_Range;
             private HpBlockStreamReader m_Reader;
-            private readonly NativeSlice<RigidBody> m_Bodies;
             public TriggerEvent Current { get; private set; }
 
-            internal Enumerator(HpLinkedRange* range, NativeSlice<RigidBody> bodies)
+            internal Enumerator(HpLinkedRange* range)
             {
                 m_Range = range;
                 m_Reader = new HpBlockStreamReader(m_Range);
                 Current = default;
-
-                m_Bodies = bodies;
             }
 
             public bool MoveNext()
@@ -52,7 +45,7 @@ namespace Unity.Physics
                 {
                     var eventData = (TriggerEventData*)m_Reader.Peek();
 
-                    Current = eventData->CreateTriggerEvent(m_Bodies);
+                    Current = eventData->CreateTriggerEvent();
 
                     m_Reader.Advance(sizeof(TriggerEventData));
                     return true;